requirements.txt
Flask
Flask-WTF
Flask-Talisman
Flask-SQLAlchemy
Flask-Login
argon2-cffi
pydantic
email-validator
